bridge/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ur5e-mcp-bridge"
version = "0.1.0"
description = "Local stdio MCP bridge for guarded Universal Robots UR5e control"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "mcp[cli]>=1.0.0",
  "pydantic>=2.6,<3",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
ur5e-mcp = "ur5e_bridge.server:main"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["../tests"]

[tool.setuptools]
packages = ["ur5e_bridge"]
